+++ examples/apache-flink/src/test/java/nl/basjes/parse/httpdlog/flink/avro/TestParserMapFunctionAvroClass.java
@@ -26,10 +26,11 @@ import nl.basjes.parse.httpdlog.dissectors.geoip.GeoIPISPDissector;
 import nl.basjes.parse.httpdlog.flink.TestCase;
 import nl.basjes.parse.webevents.Click;
 import org.apache.commons.lang3.builder.Builder;
+import org.apache.flink.api.common.functions.OpenContext;
 import org.apache.flink.api.common.functions.RichMapFunction;
-import org.apache.flink.api.java.DataSet;
-import org.apache.flink.api.java.ExecutionEnvironment;
-import org.apache.flink.configuration.Configuration;
+import org.apache.flink.streaming.api.datastream.DataStream;
+import org.apache.flink.streaming.api.environment.LocalStreamEnvironment;
+import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.junit.jupiter.api.Test;
 
 import java.io.Serializable;
@@ -82,7 +83,7 @@ class TestParserMapFunctionAvroClass implements Serializable {
         private Parser<ClickSetter> parser;
 
         @Override
-        public void open(Configuration parameters) {
+        public void open(OpenContext openContext) {
             parser = new HttpdLoglineParser<>(ClickSetter.class, TestCase.getLogFormat())
                 .addDissector(new ScreenResolutionDissector())
                 .addTypeRemapping("request.firstline.uri.query.g", "HTTP.URI")
@@ -106,20 +107,20 @@ class TestParserMapFunctionAvroClass implements Serializable {
     @Test
     void testClassDefinitionAvro() throws Exception {
         // set up the execution environment
-        final ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
+        final StreamExecutionEnvironment env = LocalStreamEnvironment.getExecutionEnvironment();
+        env.setParallelism(1);
+        DataStream<String> input = env.fromData(TestCase.getInputLine());
 
-        DataSet<String> input = env.fromElements(TestCase.getInputLine());
-
-        DataSet<Click> filledTestRecords = input
+        DataStream<Click> filledTestRecords = input
             .map(new MyParserMapper())
             .name("Extract Elements from logline");
 
         filledTestRecords.print();
 
-        List<Click> result = filledTestRecords.collect();
+        List<Click> result = filledTestRecords.executeAndCollect(100);
 
         assertEquals(1, result.size());
-        assertEquals(ExpectedClick.create(), result.get(0));
+        assertEquals(ExpectedClick.create(), result.getFirst());
     }
 
 }