+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v5.12.0-SNAPSHOT
 ===
 - Updated to need Java 21
 - Update Serde to Hive 4.2.0
+- Updated Flink and Beam examples to latest versions
 - GeoIP update no longer supports the "metrocode"
 
 v5.11.0